+++ lib/annotation.dart
@@ -1,0 +1,13 @@
+// deprecated. removed for Flutter support.
+
+// /// Route Annontation Model Class
+// class Route {
+//   /// Method for the request (e.g. GET, POST, PUT, DELETE)
+//   final String method;
+
+//   /// The path for the request (e.g. /, /posts, /api/posts)
+//   final String path;
+
+//   /// Constructor
+//   const Route(this.method, this.path);
+// }