+++ lib/config.dart
@@ -1,6 +1,9 @@
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:yaml/yaml.dart';
+import 'package:toml/toml.dart';
+
 /// Config class
 class Config {
   final Map<String, dynamic> _settings = {};
@@ -14,9 +17,19 @@ class Config {
     final file = File(filePath);
     if (file.existsSync()) {
       final jsonString = file.readAsStringSync();
-      _settings.addAll(jsonDecode(jsonString));
+      if (filePath.endsWith('.json')) {
+        _settings.addAll(jsonDecode(jsonString));
+      } else if (filePath.endsWith('.yaml')) {
+        _settings.addAll(loadYaml(jsonString));
+      } else if (filePath.endsWith('.toml')) {
+        var config = TomlDocument.parse(jsonString).toMap();
+        _settings.addAll(config);
+      } else {
+        print('[dartcore] Invalid configuration file format: $filePath');
+      }
     } else {
-      print('[dartcore] Configuration file not found: $filePath');
+      print(
+          '[dartcore] Configuration file not found: ${filePath.replaceFirst("", "null")}');
     }
   }