+++ lib/dartcore.dart
@@ -1,19 +1,26 @@
+// ignore_for_file: unnecessary_null_comparison
+
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'package:dartcore/apikeymanager.dart';
+import 'package:dartcore/blocker.dart';
 import 'package:dartcore/cache.dart';
 import 'package:dartcore/config.dart';
+import 'package:dartcore/custom_types.dart';
 import 'package:dartcore/event_emitter.dart';
+import 'package:dartcore/openapi_spec.dart';
 import 'package:dartcore/rate_limiter.dart';
 import 'package:dartcore/template_engine.dart';
+import 'package:dio/dio.dart' as dio;
+import 'package:http/http.dart' as http;
 import 'package:mime/mime.dart';
 
 /// dartcore's version
-const version = '0.0.5';
+const version = '0.0.7';
 
 /// Defines Handler
-typedef Handler = Future<void> Function(HttpRequest request);
+typedef Handler = Future<void> Function(HttpRequest request, Response response);
 
 /// Defines Middleware
 typedef Middleware = Future<void> Function(HttpRequest request, Function next);
@@ -26,20 +33,29 @@ class App {
   TemplateEngine? _templateEngine;
   final Cache _cache = Cache();
   final Map<String, Map<String, Handler>> _routes = {};
+  Map<String, dynamic> _metadata = {};
   final List<Middleware> _middlewares = [];
   Function(HttpRequest request)? _custom404;
   Function(HttpRequest request, Object error)? _custom500;
 
+  /// Debug mode (TODO)
+  final bool debug = true;
+
   RateLimiter? _rateLimiter;
 
   /// Constructor
-  App(String configPath) {
-    _config = Config(configPath);
+  App({String? configPath, bool? debug, Map<String, dynamic>? metadata}) {
+    _config = Config(configPath ?? "");
+    _metadata = metadata ??
+        {"title": "My API", "version": "1.0.0", "description": "Dartcore API"};
     _templateEngine = TemplateEngine();
     _rateLimiter = RateLimiter(
+      shouldDisplayCaptcha: false,
       storagePath: './rate_limits.dartcorelimits',
       maxRequests: 60,
       resetDuration: Duration(minutes: 1),
+      ipBlocker: IPBlocker(), // Yeah, recommend changing it
+      countryBlocker: CountryBlocker(), // Yeah...
       encryptionPassword:
           "dartcore", // Don't keep it unless it's a private server, recommend changing it
     );
@@ -55,28 +71,169 @@ class App {
     return version;
   }
 
-  /// Routing function
-
-  void route(String method, String path, Handler handler) {
-    _routes.putIfAbsent(method, () => {})[path] = handler;
+  /// Returns the routes
+  /// e.g. final routes = app.routes();
+  Map<String, Map<String, Handler>> routes() {
+    return _routes;
   }
 
-  /// Using middleware function
+  /// Returns all the metadata
+  /// e.g. final metadata = app.metadata();
+  Map<String, dynamic> metadata() {
+    return _metadata;
+  }
 
+  /// Adds a route to the application with the specified HTTP method and path.
+  /// Associates the route with a handler function and optional metadata.
+  ///
+  /// - Parameters:
+  ///   - method: The HTTP method for the route (e.g., 'GET', 'POST').
+  ///   - path: The path for the route (e.g., '/api/resource').
+  ///   - handler: The function to handle requests to this route.
+  ///   - metadata: Optional metadata for the route, such as description or status.
+  void route(String method, String path, Handler handler,
+      {Map<String, dynamic>? metadata}) {
+    _routes.putIfAbsent(method, () => {})[path] = handler;
+    if (metadata != null) {
+      _metadata.putIfAbsent(path, () => {})[method] = metadata;
+    }
+  }
+
+  /// Adds a GET route to the application with the specified path.
+  /// Associates the route with a handler function and optional metadata.
+  ///
+  /// - Parameters:
+  ///   - path: The path for the route (e.g., '/api/resource').
+  ///   - handler: The function to handle requests to this route.
+  ///   - metadata: Optional metadata for the route, such as description or status.
+  void get(String path, Handler handler, {Map<String, dynamic>? metadata}) {
+    final method = 'GET';
+    _routes.putIfAbsent(method, () => {})[path] = handler;
+    if (metadata != null) {
+      _metadata.putIfAbsent(path, () => {})[method] = metadata;
+    }
+  }
+
+  /// Adds a POST route to the application with the specified path.
+  /// Associates the route with a handler function and optional metadata.
+  ///
+  /// - Parameters:
+  ///   - path: The path for the route (e.g., '/api/resource').
+  ///   - handler: The function to handle requests to this route.
+  ///   - metadata: Optional metadata for the route, such as description or status.
+  void post(String path, Handler handler, {Map<String, dynamic>? metadata}) {
+    final method = 'POST';
+    _routes.putIfAbsent(method, () => {})[path] = handler;
+    if (metadata != null) {
+      _metadata.putIfAbsent(path, () => {})[method] = metadata;
+    }
+  }
+
+  /// Adds a PUT route to the application with the specified path.
+  /// Associates the route with a handler function and optional metadata.
+  ///
+  /// - Parameters:
+  ///   - path: The path for the route (e.g., '/api/resource').
+  ///   - handler: The function to handle requests to this route.
+  ///   - metadata: Optional metadata for the route, such as description or status.
+  void put(String path, Handler handler, {Map<String, dynamic>? metadata}) {
+    final method = 'PUT';
+    _routes.putIfAbsent(method, () => {})[path] = handler;
+    if (metadata != null) {
+      _metadata.putIfAbsent(path, () => {})[method] = metadata;
+    }
+  }
+
+  /// Adds a DELETE route to the application with the specified path.
+  /// Associates the route with a handler function and optional metadata.
+  ///
+  /// - Parameters:
+  ///   - path: The path for the route (e.g., '/api/resource').
+  ///   - handler: The function to handle requests to this route.
+  ///   - metadata: Optional metadata for the route, such as description or status.
+  void delete(String path, Handler handler, {Map<String, dynamic>? metadata}) {
+    final method = 'DELETE';
+    _routes.putIfAbsent(method, () => {})[path] = handler;
+    if (metadata != null) {
+      _metadata.putIfAbsent(path, () => {})[method] = metadata;
+    }
+  }
+
+  /// Converts an IP to a Location
+  Future<String?> getGeoLocation(String ip) async {
+    final url = 'http://ip-api.com/json/$ip';
+    final response = await http.get(Uri.parse(url));
+
+    if (response.statusCode == 200) {
+      final data = jsonDecode(response.body);
+      return data['country'];
+    }
+    return null;
+  }
+
+  /// Checks if a country is blocked
+  Future<bool> isBlockedCountry(String ip) async {
+    final country = await getGeoLocation(ip);
+    final blockedCountries = _rateLimiter!.countryBlocker.blockedList;
+
+    return country != null && blockedCountries.contains(country);
+  }
+
+  /// Adds a middleware to the list of middlewares that will be executed in order.
+  /// Note that the order of execution is the same as the order of addition.
+  /// You can add a middleware at any time, but it will only start executing on new requests.
+  /// If you add a middleware after the server is started, it will be executed on new requests,
+  /// but not on requests that are already in progress.
   void use(Middleware middleware) {
     _middlewares.add(middleware);
   }
 
-  /// Starts the server
-
+  /// Starts the server.
+  ///
+  /// If [port] is not provided, it will use 8080.
+  /// If [port] is `0`, it will use a random port chosen by the OS.
+  /// If [address] is not provided, it will use `localhost`, `127.0.0.1`, `0.0.0.0` or `::1`
+  ///
+  /// If the server fails to start, it will print an error message and exit with code 1.
+  ///
+  /// If the server starts successfully, it will print a success message, and the server
+  /// will listen for incoming requests.
+  ///
+  /// If the server is interrupted with Ctrl+C, it will shutdown the server, delete the
+  /// rate limit file if it exists, and exit with code 0.
   Future<void> start({String address = '0.0.0.0', int port = 8080}) async {
-    _server = await HttpServer.bind(address, port);
+    try {
+      _server = await HttpServer.bind(address, port);
+    } on SocketException catch (e) {
+      if (e.osError!.errorCode == 10048) {
+        print(
+            '[dartcore] Port is already in use, try to use to a different port.');
+        exit(1);
+      } else {
+        print(
+            '[dartcore] Failed to start server: OS ERROR ${e.osError!.errorCode}, ${e.osError!.message}');
+        exit(1);
+      }
+    } catch (e) {
+      print('[dartcore] Failed to start server:');
+      print(e);
+      exit(1);
+    }
     print(
-        '[dartcore] Server running on ${address.replaceFirst('0.0.0.0', 'All IP Addresses on port ')}:$port');
+        '[dartcore] Server running on ${_server!.address.address}:${_server!.port}\n[dartcore] Press Ctrl+C to shutdown the Server.');
     emit('serverStarted', {'address': address, 'port': port});
 
+    ProcessSignal.sigint.watch().listen((_) async {
+      final file = File(_rateLimiter!.storagePath);
+      if (await file.exists()) {
+        await file.delete();
+      }
+      await shutdown();
+      exit(0);
+    });
+
     await for (HttpRequest request in _server!) {
-      await _handleRequest(request);
+      await _handleRequest(request, Response(request: request));
     }
   }
 
@@ -85,7 +242,51 @@ class App {
     return _config?.get(key);
   }
 
-  Future<void> _handleRequest(HttpRequest request) async {
+  /// Makes SwaggerUI routes (can be copy-pasted and edited but just add `app.` before each `route()`)
+  void openApi() {
+    route("GET", "/openapi", (req, res) async {
+      res.json(generateOpenApiSpec(this,
+          true)); // set it to `false` in case you are copy-pasting and editing the names of the routes
+    });
+    route("GET", "/docs", (req, res) async {
+      await res.staticFile("./swaggerui/index.html", ContentType.html);
+    });
+    route("GET", "/index.css", (req, res) async {
+      await res.staticFile("./swaggerui/index.css", ContentType("text", "css"));
+    });
+    route("GET", "/swagger-ui-bundle.js", (req, res) async {
+      await res.send(
+          (await http.get(Uri.parse(
+                  "https://raw.githubusercontent.com/swagger-api/swagger-ui/refs/heads/master/dist/swagger-ui-bundle.js")))
+              .body,
+          ContentType("application", "javascript"));
+    });
+    route("GET", "/swagger-initializer.js", (req, res) async {
+      await res.staticFile("./swaggerui/swagger-initializer.js",
+          ContentType("application", "javascript"));
+    });
+    route("GET", "/swagger-ui-standalone-preset.js", (req, res) async {
+      await res.send(
+          (await dio.Dio().get(
+                  "https://raw.githubusercontent.com/swagger-api/swagger-ui/refs/heads/master/dist/swagger-ui-standalone-preset.js",
+                  options: dio.Options(
+                    responseType: dio.ResponseType.json,
+                    followRedirects: true,
+                    headers: {
+                      'Content-Type': 'application/json; charset=utf-8',
+                    },
+                    contentType: 'application/json; charset=utf-8',
+                  )))
+              .data,
+          ContentType("application", "javascript"));
+    });
+    route("GET", "/swagger-ui.css", (req, res) async {
+      await res.staticFile(
+          "./swaggerui/swagger-ui.css", ContentType("text", "css"));
+    });
+  }
+
+  Future<void> _handleRequest(HttpRequest request, Response response) async {
     final path = request.uri.path;
     final method = request.method;
 
@@ -94,14 +295,21 @@ class App {
     await _runMiddlewares(request, () async {
       final clientIp =
           request.connectionInfo?.remoteAddress.address ?? 'unknown';
-      if (_rateLimiter!.isRateLimited(clientIp)) {
+      if (await _rateLimiter!.isRateLimited(clientIp)) {
         _handleRateLimit(request);
+        return;
+      }
+      if (await isBlockedCountry(clientIp)) {
+        request.response.statusCode = HttpStatus.forbidden;
+        request.response.write(
+            'Access from your country is blocked.\n[dartcore v$version]');
+        await request.response.close();
         return;
       }
 
       final handler = _findHandler(method, path);
       if (handler != null) {
-        await handler(request);
+        await handler(request, response);
         print('[dartcore] $method $path --> ${request.response.statusCode}');
       } else {
         _custom404?.call(request) ?? _handle404(request);
@@ -138,10 +346,42 @@ class App {
   }
 
   void _handleRateLimit(HttpRequest request) {
-    request.response
-      ..statusCode = HttpStatus.tooManyRequests
-      ..write('Rate limit exceeded\n[dartcore v$version]')
-      ..close();
+    if (!_rateLimiter!.shouldDisplayCaptcha) {
+      request.response
+        ..statusCode = HttpStatus.tooManyRequests
+        ..write('Rate limit exceeded\n[dartcore v$version]')
+        ..close();
+    } else {
+      final res = Response(request: request);
+      res.html("""
+<html>
+<title>Rate Limit Exceeded - Dartcore</title>
+    <body>
+      <p>You have exceeded the maximum number of requests. Please complete the CAPTCHA to continue:</p>
+      <div class="g-recaptcha" data-sitekey="YOUR_SITE_KEY"></div>
+      <script src="https://www.google.com/recaptcha/api.js" async defer></script>
+      <form action="/verify-captcha" method="post">
+        <input type="hidden" name="captchaResponse" value="CAPTCHA_RESPONSE">
+        <button type="submit">Submit</button>
+      </form>
+    </body>
+  </html>
+""");
+    }
+  }
+
+  /// Verifies the CAPTCHA response
+  Future<bool> verifyCaptcha(String response, String secretKey) async {
+    final url = Uri.parse('https://www.google.com/recaptcha/api/siteverify');
+    final responseH = await http.post(
+      url,
+      body: {
+        'secret': secretKey,
+        'response': response,
+      },
+    );
+    final responseBody = json.decode(responseH.body);
+    return responseBody['success'] == true;
   }
 
   Handler? _findHandler(String method, String path) {
@@ -187,20 +427,6 @@ class App {
     _custom500 = handler;
   }
 
-  /// Sends a json response
-  Future<void> sendJson(HttpRequest request, Map<String, dynamic> data) async {
-    request.response
-      ..headers.contentType = ContentType.json
-      ..write(jsonEncode(data))
-      ..close();
-  }
-
-  /// Sets a header to the response
-
-  void setHeader(HttpRequest request, String key, String value) {
-    request.response.headers.set(key, value);
-  }
-
   void _handle500(HttpRequest request, Object error) {
     _custom500 != null ? _custom500!(request, error) : request.response
       ..statusCode = HttpStatus.internalServerError
@@ -209,10 +435,15 @@ class App {
     print('[dartcore] ${request.method} ${request.uri.path} --> 500: $error');
   }
 
-  /// Serves a static file
-
+  /// Serves a static file at [filePath] to the client.
+  ///
+  /// The file is sent to the client with the correct MIME type.
+  ///
+  /// If the file does not exist, a 404 error is sent to the client and
+  /// the request is not closed.
   Future<void> serveStaticFile(HttpRequest request, String filePath) async {
     final file = File(filePath);
+
     if (await file.exists()) {
       final fileStream = file.openRead();
       await fileStream.pipe(request.response);
@@ -221,21 +452,40 @@ class App {
     }
   }
 
-  /// Parses a JSON from the request body
-
+  /// Parses the request body as JSON and returns it as a Map<String, dynamic>
+  ///
+  /// The request body should contain a valid JSON string.
+  ///
+  /// Throws a FormatException if the request body is not a valid JSON string.
   Future<Map<String, dynamic>> parseJson(HttpRequest request) async {
-    final content = await utf8.decoder.bind(request).join();
-    return jsonDecode(content) as Map<String, dynamic>;
+    final content = await utf8.decodeStream(request);
+    return jsonDecode(content);
   }
 
-  /// Gets query parameters (e.g. getQueryParams(request)['name'])
+  /// Retrieves query parameters from the given HTTP request.
+  ///
+  /// Returns a map containing key-value pairs of query parameters.
+  /// For example, a request with the URI `/endpoint?name=value` will
+  /// return `{'name': 'value'}`.
+  ///
+  /// - Parameter request: The HTTP request containing the URI with query parameters.
 
-  Map<String, String> getQueryParams(HttpRequest request) {
+  Map<String, String> params(HttpRequest request) {
     return request.uri.queryParameters;
   }
 
-  /// Parses Multipart Requests, useful for adding upload a file
-
+  /// Parses a multipart/form-data request and saves the uploaded files
+  ///
+  /// to the [saveTo] directory.
+  ///
+  /// The request body should contain a valid multipart/form-data content.
+  ///
+  /// Throws a 500 error if the request body is not a valid multipart/form-data
+  /// content.
+  ///
+  /// - [request]: The HTTP request containing the multipart/form-data
+  ///   content.
+  /// - [saveTo]: The directory where the uploaded files will be saved.
   Future<void> parseMultipartRequest(HttpRequest request, String saveTo) async {
     if (request.headers.contentType?.mimeType == 'multipart/form-data') {
       final boundary = request.headers.contentType?.parameters['boundary'];
@@ -263,36 +513,6 @@ class App {
     }
   }
 
-  /// Serves a file, useful for downloading a file
-
-  Future<void> sendFile(HttpRequest request, File file) async {
-    if (await file.exists()) {
-      request.response.headers.contentType = ContentType.binary;
-      await file.openRead().pipe(request.response);
-    } else {
-      _handle404(request);
-    }
-  }
-
-  /// Sends an HTML data
-
-  Future<void> sendHtml(HttpRequest request, String htmlContent) async {
-    request.response
-      ..headers.contentType = ContentType.html
-      ..write(htmlContent)
-      ..close();
-  }
-
-  /// Sends a data with a custom type
-
-  Future<void> send(
-      HttpRequest request, String content, ContentType contentType) async {
-    request.response
-      ..headers.contentType = contentType
-      ..write(content)
-      ..close();
-  }
-
   /// Renders a template and sends it as a response
   Future<void> renderTemplate(HttpRequest request, String templatePath,
       Map<String, dynamic> context) async {
@@ -312,7 +532,7 @@ class App {
   Middleware errorHandlingMiddleware() {
     return (HttpRequest request, Function next) async {
       try {
-        await next(); // Continue to the next middleware or route handler
+        await next();
       } catch (e, stackTrace) {
         _handleError(request, e, stackTrace);
       }
@@ -352,7 +572,7 @@ class App {
   Future<void> shutdown() async {
     if (_server != null) {
       await _server!.close();
-      print('[dartcore] Server shutdown.');
+      print('[dartcore] Server shutting down.');
       emit('serverShutdown', {'message': 'Server has been shut down.'});
     }
   }
@@ -374,18 +594,23 @@ class App {
     };
   }
 
-  /// Sets up a basic API key management routes
-
-  void setupApiKeyRoutes(App app, ApiKeyManager apiKeyManager) {
-    app.route('POST', '/api-keys', (HttpRequest request) async {
+  /// Sets up the following routes for basic API key management:
+  ///
+  /// - `POST /api-keys`: Generates a new API key and returns it in the response
+  ///   body.
+  /// - `DELETE /api-keys/:key`: Revokes the API key with the given `:key`.
+  ///
+  /// The API keys are managed by the given [apiKeyManager].
+  void setupApiKeyRoutes(ApiKeyManager apiKeyManager) {
+    route('POST', '/api-keys', (req, res) async {
       final newApiKey = apiKeyManager.generateApiKey();
-      await app.sendJson(request, {'apiKey': newApiKey.key});
+      await res.json({'apiKey': newApiKey.key});
     });
 
-    app.route('DELETE', '/api-keys/:key', (HttpRequest request) async {
-      final key = request.uri.pathSegments.last;
+    route('DELETE', '/api-keys/:key', (req, res) async {
+      final key = req.uri.pathSegments.last;
       apiKeyManager.revokeApiKey(key);
-      request.response
+      req.response
         ..statusCode = HttpStatus.noContent
         ..close();
     });