+++ CHANGELOG.md
@@ -34,3 +34,22 @@
 - Added `app.shutdown();` to gracefully shutdown the server
 - Added Caching for responses
 - Now RateLimter uses Storage instead of memory, also encrypts it with a password
+
+## 0.0.6
+
+- New Syntax!   old: app.route("GET","/",(req) {});  new: app.route("GET","/",(req,res) {});
+- YAML Support for Configs
+- TOML Support for Configs
+- Annotation Support!!! See example for more details!
+- Database integration! (Currently just MySQL and PostgreSQL)
+- OpenAPI Auto-generation + SwaggerUI!!
+- Improved encrypting for Rate limits
+- Added IP Blocking :)
+- Added Geo-blocking :)
+- Added Captcha for Rate-limit (TODO)
+- Fixed Rate Limits crash on second startup
+- Added Shutdown-On-SIGINT (CTRL+C shuts down the server)
+
+## 0.0.7
+- Removed Annotation for Flutter and AOT support
+- Better documentation for some methods